+++ modules/tinymce/src/themes/silver/test/ts/headless/window/SilverDialogReuseTest.ts
@@ -1,5 +1,4 @@
-import { ApproxStructure, Assertions, FocusTools, StructAssert, UiFinder } from '@ephox/agar';
-import { TestHelpers } from '@ephox/alloy';
+import { ApproxStructure, Assertions, FocusTools, StructAssert, TestStore, UiFinder } from '@ephox/agar';
 import { afterEach, before, describe, it } from '@ephox/bedrock-client';
 import { Dialog as BridgeSpec } from '@ephox/bridge';
 import { Optional } from '@ephox/katamari';
@@ -13,7 +12,7 @@ import * as WindowManager from 'tinymce/themes/silver/ui/dialog/WindowManager';
 import * as TestExtras from '../../module/TestExtras';
 
 describe('headless.tinymce.themes.silver.window.SilverDialogReuseTest', () => {
-  const store = TestHelpers.TestStore();
+  const store = TestStore();
   const helpers = TestExtras.bddSetup();
   let windowManager: WindowManagerImpl;
   let dialogApi: Dialog.DialogInstanceApi<any>;