+++ modules/tinymce/src/themes/silver/test/ts/module/DialogUtils.ts
@@ -1,5 +1,4 @@
-import { UiFinder } from '@ephox/agar';
-import { TestHelpers } from '@ephox/alloy';
+import { TestStore, UiFinder } from '@ephox/agar';
 import { SugarBody } from '@ephox/sugar';
 import { TinyUiActions } from '@ephox/wrap-mcagar';
 
@@ -10,7 +9,7 @@ import { WindowParams } from 'tinymce/core/api/WindowManager';
 const open = <T>(editor: Editor, spec: Dialog.DialogSpec<T>, params: WindowParams): Dialog.DialogInstanceApi<T> =>
   editor.windowManager.open(spec, params);
 
-const openWithStore = <T>(editor: Editor, spec: Dialog.DialogSpec<T>, params: WindowParams, store: TestHelpers.TestStore): Dialog.DialogInstanceApi<T> => {
+const openWithStore = <T>(editor: Editor, spec: Dialog.DialogSpec<T>, params: WindowParams, store: TestStore): Dialog.DialogInstanceApi<T> => {
   const dialogSpec = {
     onSubmit: store.adder('onSubmit'),
     onClose: store.adder('onClose'),